Small reorganization
