--- src/ros2/src/deprecated/bob_simulator/README.md
@@ -1,15 +1,0 @@
-# Simulator Package
-
-## Overview
-
-### Description: 
-
-A package providing a number of randomly moving objects simulation and can be used to overlay onto either a video or rtsp camera source so as to test the system.
-
-### Vision: 
-
-
-## Nodes:
-
-- Object Simulator Node
-- Simulation Overlay Provider Node
