--- src/ros2/src/deprecated/bob_simulator/package.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-    <name>bob_simulator</name>
-    <version>1.2.16</version>
-    <description>ROS2 Node for Simulating and Publishing Moving Objects on a Frame</description>
-    <maintainer email="uapbob@gmail.com">Bob Universal Object Tracker</maintainer>
-    <license>GPLv3</license>
-
-    <buildtool_depend>ament_cmake</buildtool_depend>
-
-    <depend>rclcpp</depend>
-    <depend>sensor_msgs</depend>
-    <depend>std_msgs</depend>
-    <depend>cv_bridge</depend>
-    <depend>rcl_interfaces</depend>
-    <depend>rclcpp_components</depend>
-
-    <exec_depend>bob_shared</exec_depend>
-
-    <export>
-        <build_type>ament_cmake</build_type>
-    </export>
-</package>
